small refactoring of encoding and test
