+++ encoding.go
@@ -3,7 +3,18 @@ package katzenmint
 import (
 	"encoding/hex"
 	"fmt"
+
+	"github.com/ugorji/go/codec"
 )
+
+var jsonHandle *codec.JsonHandle
+
+func init() {
+	jsonHandle = new(codec.JsonHandle)
+	jsonHandle.Canonical = true
+	jsonHandle.IntegerAsString = 'A'
+	jsonHandle.MapKeyAsString = true
+}
 
 // DecodeHex return byte of the given hex string
 // return nil if the src is not valid hex string
@@ -23,5 +34,16 @@ func DecodeHex(src string) (out []byte) {
 // bytes
 func EncodeHex(src []byte) (out string) {
 	out = hex.EncodeToString(src)
+	return
+}
+
+func DecodeJson(data []byte, v interface{}) error {
+	dec := codec.NewDecoderBytes(data, jsonHandle)
+	err := dec.Decode(v)
+	return err
+}
+
+func EncodeJson(v interface{}) (data []byte, err error) {
+	err = codec.NewEncoderBytes(&data, jsonHandle).Encode(v)
 	return
 }
